requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
h11==0.16.0
httptools==0.6.4
idna==3.10
MouseInfo==0.1.3
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pydantic==2.11.3
pydantic_core==2.33.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dotenv==1.1.0
python-multipart==0.0.20
pytweening==1.2.0
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
